Исправить причину ошибки тестов
