--- src/App/router/index.ts
@@ -1,2 +1,0 @@
-export { AppRouter } from "./AppRouter";
-export { routesPaths } from "./routesPaths";
