+++ src/app/pages/Contacts.tsx
@@ -1,14 +1,14 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { routesPaths } from "../router/routesPaths";
+import { routesPaths } from "../router";
 
-const About: React.FC = () => {
+const Contacts: React.FC = () => {
   return (
     <div>
-      <h1>О нас</h1>
+      <h1>Контакты</h1>
       <Link to={routesPaths.home}>На главную</Link>
     </div>
   );
 };
 
-export default About;
+export default Contacts;
