+++ src/app/pages/ProjectsList.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { routesPaths } from "../router/routesPaths";
+import { routesPaths } from "../router";
 
 const ProjectsList: React.FC = () => {
   return (
