+++ src/app/router/index.ts
@@ -1,0 +1,2 @@
+export { AppRouter } from "./AppRouter";
+export { routesPaths } from "./routesPaths"
